pcre32: compile: Fix signed/unsigned mismatch in adjust_recurse
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1073 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_compile.c
@@ -2810,7 +2810,7 @@ while ((ptr = (pcre_uchar *)find_recurse(ptr, utf)) != NULL)
 
   for (hc = save_hwm; hc < cd->hwm; hc += LINK_SIZE)
     {
-    offset = GET(hc, 0);
+    offset = (int)GET(hc, 0);
     if (cd->start_code + offset == ptr + 1)
       {
       PUT(hc, 0, offset + adjust);
@@ -2823,7 +2823,7 @@ while ((ptr = (pcre_uchar *)find_recurse(ptr, utf)) != NULL)
 
   if (hc >= cd->hwm)
     {
-    offset = GET(ptr, 1);
+    offset = (int)GET(ptr, 1);
     if (cd->start_code + offset >= group) PUT(ptr, 1, offset + adjust);
     }
